Feature/template

See merge request sedordiga/bariHosh!4

+++ src/main/java/br/com/bariHosh/dao/GenericDAO.java
@@ -1,4 +1,5 @@
 package br.com.bariHosh.dao;
+
 import java.util.List;
 
 public interface GenericDAO<T> {
@@ -8,9 +9,7 @@ public interface GenericDAO<T> {
 
 	public void excluir(T model);
 
-	public T carregar(Long id);
-
-//	public T buscarPorLogin(String login);
+	public T carregar(Integer id);
 
 	public List<T> listar();
 }
